Split components into their own files. CSS is now working
